Remove dynamic namespace and replace old code
