--- src/dynamic/mod.rs
@@ -1,11 +1,0 @@
-//! Utilities for dynamic application helpers.
-
-pub(crate) mod app;
-pub(crate) mod state;
-pub(crate) mod trans;
-
-pub use self::{
-    app::{Application, ApplicationBuilder, CoreApplication, CoreApplicationBuilder},
-    state::{GlobalCallback, State, StateCallback, StateError, StateMachine, StateStorage, States},
-    trans::{Trans, TransEvent},
-};
