+++ src/state.rs
@@ -1,11 +1,48 @@
-//! Utilities for game state management.
+//! Dynamic utilities for game state management.
 
-use amethyst_input::is_close_requested;
+use crate::{ecs::prelude::World, error::Error, Trans};
 
-use crate::{ecs::prelude::World, game_data::GameData, StateEvent};
+use hashbrown;
+use smallvec::SmallVec;
+use std::fmt;
+use std::hash;
+use std::mem;
 
-use std::fmt::Result as FmtResult;
-use std::fmt::{Display, Formatter};
+/// The type of a new state.
+pub(crate) type NewState<S, E> = (S, Box<dyn StateCallback<S, E> + Send + Sync>);
+
+/// A resource used to indirectly manipulate the contents of the state machine.
+pub struct States<S, E> {
+    new_states: SmallVec<[NewState<S, E>; 16]>,
+}
+
+impl<S, E> Default for States<S, E> {
+    fn default() -> Self {
+        States {
+            new_states: Default::default(),
+        }
+    }
+}
+
+impl<S, E> States<S, E> {
+    /// Indicate that we want to create a new state.
+    pub fn new_state<C>(&mut self, state: S, callback: C)
+    where
+        C: 'static + StateCallback<S, E> + Send + Sync,
+    {
+        self.new_states.push((state, Box::new(callback)));
+    }
+
+    /// Drain all new states and push into the provided callback.
+    pub fn drain_new_states<C>(&mut self, mut c: C)
+    where
+        C: FnMut(NewState<S, E>),
+    {
+        for new_state in self.new_states.drain() {
+            c(new_state)
+        }
+    }
+}
 
 /// Error type for errors occurring in StateMachine
 #[derive(Debug)]
@@ -16,8 +53,8 @@ pub enum StateError {
     CallbackConflict,
 }
 
-impl Display for StateError {
-    fn fmt(&self, fmt: &mut Formatter<'_>) -> FmtResult {
+impl fmt::Display for StateError {
+    fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         match *self {
             StateError::NoStatesPresent => write!(
                 fmt,
@@ -30,89 +67,96 @@ impl Display for StateError {
     }
 }
 
-/// State data encapsulates the data sent to all state functions from the application main loop.
-pub struct StateData<'a, T>
+/// The trait associated with a stage.
+pub trait State<E>: Clone + Default + fmt::Debug
 where
-    T: 'a,
+    Self: Sized,
 {
-    /// Main `World`
-    pub world: &'a mut World,
-    /// User defined game data
-    pub data: &'a mut T,
+    /// The storage used for storing callbacks for the given state.
+    type Storage: Default + StateStorage<Self, E>;
 }
 
-impl<'a, T> StateData<'a, T>
+/// Provides access to storage for states.
+pub trait StateStorage<S, E>
 where
-    T: 'a,
+    Self: Sized,
 {
-    /// Create a new state data
-    pub fn new(world: &'a mut World, data: &'a mut T) -> Self {
-        StateData { world, data }
+    /// Insert the given callback, returning an existing callback if it is already present.
+    fn insert(
+        &mut self,
+        state: S,
+        callback: Box<dyn StateCallback<S, E>>,
+    ) -> Option<Box<dyn StateCallback<S, E>>>;
+
+    /// Get mutable storage for the given state.
+    fn get_mut(&mut self, value: &S) -> Option<&mut Box<dyn StateCallback<S, E>>>;
+
+    /// Apply the specified function to all values.
+    fn do_values<F>(&mut self, apply: F)
+    where
+        F: FnMut(&mut Box<dyn StateCallback<S, E>>);
+}
+
+/// A callback that is registered for all events.
+/// This is typically used for bookkeeping specific things.
+pub trait GlobalCallback<S, E> {
+    /// Fired when state machine has been started.
+    fn started(&mut self, _: &mut World) {}
+
+    /// Fired when state machine has been stopped.
+    fn stopped(&mut self, _: &mut World) {}
+
+    /// Fired when state has changed, and what it was changed to.
+    fn changed(&mut self, _: &mut World, _: &S) {}
+
+    /// Fired on events.
+    ///
+    /// If multiple callbacks would result in a state transition, they will be applied one after
+    /// another in an undetermined order.
+    fn handle_event(&mut self, _: &mut World, _: &E) -> Trans<S> {
+        Trans::None
+    }
+
+    /// Fired on fixed updates.
+    fn fixed_update(&mut self, _: &mut World) -> Trans<S> {
+        Trans::None
+    }
+
+    /// Fired on updates.
+    fn update(&mut self, _: &mut World) -> Trans<S> {
+        Trans::None
     }
 }
 
-/// Types of state transitions.
-/// T is the type of shared data between states.
-/// E is the type of events
-pub enum Trans<T, E> {
-    /// Continue as normal.
-    None,
-    /// Remove the active state and resume the next state on the stack or stop
-    /// if there are none.
-    Pop,
-    /// Pause the active state and push a new state onto the stack.
-    Push(Box<dyn State<T, E>>),
-    /// Remove the current state on the stack and insert a different one.
-    Switch(Box<dyn State<T, E>>),
-    /// Stop and remove all states and shut down the engine.
-    Quit,
-}
-
-/// Event queue to trigger state `Trans` from other places than a `State`'s methods.
-/// # Example:
-/// ```rust, ignore
-/// world.write_resource::<EventChannel<TransEvent<MyGameData, StateEvent>>>().single_write(Box::new(|| Trans::Quit));
-/// ```
-///
-/// Transitions will be executed sequentially by Amethyst's `CoreApplication` update loop.
-pub type TransEvent<T, E> = Box<dyn Fn() -> Trans<T, E> + Send + Sync + 'static>;
-
-/// An empty `Trans`. Made to be used with `EmptyState`.
-pub type EmptyTrans = Trans<(), StateEvent>;
-
-/// A simple default `Trans`. Made to be used with `SimpleState`.
-/// By default it contains a `GameData` as its `StateData` and doesn't have a custom event type.
-pub type SimpleTrans<'a, 'b> = Trans<GameData<'a, 'b>, StateEvent>;
-
-/// A trait which defines game states that can be used by the state machine.
-pub trait State<T, E: Send + Sync + 'static> {
+/// A callback that is registered for a specific state.
+pub trait StateCallback<S, E> {
     /// Executed when the game state begins.
-    fn on_start(&mut self, _data: StateData<'_, T>) {}
+    fn on_start(&mut self, _: &mut World) {}
 
     /// Executed when the game state exits.
-    fn on_stop(&mut self, _data: StateData<'_, T>) {}
+    fn on_stop(&mut self, _: &mut World) {}
 
     /// Executed when a different game state is pushed onto the stack.
-    fn on_pause(&mut self, _data: StateData<'_, T>) {}
+    fn on_pause(&mut self, _: &mut World) {}
 
     /// Executed when the application returns to this game state once again.
-    fn on_resume(&mut self, _data: StateData<'_, T>) {}
+    fn on_resume(&mut self, _: &mut World) {}
 
-    /// Executed on every frame before updating, for use in reacting to events.
-    fn handle_event(&mut self, _data: StateData<'_, T>, _event: E) -> Trans<T, E> {
+    /// Fired on events.
+    fn handle_event(&mut self, _: &mut World, _: &E) -> Trans<S> {
         Trans::None
     }
 
     /// Executed repeatedly at stable, predictable intervals (1/60th of a second
     /// by default),
     /// if this is the active state.
-    fn fixed_update(&mut self, _data: StateData<'_, T>) -> Trans<T, E> {
+    fn fixed_update(&mut self, _: &mut World) -> Trans<S> {
         Trans::None
     }
 
     /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
     /// if this is the active state.
-    fn update(&mut self, _data: StateData<'_, T>) -> Trans<T, E> {
+    fn update(&mut self, _: &mut World) -> Trans<S> {
         Trans::None
     }
 
@@ -120,432 +164,625 @@ pub trait State<T, E: Send + Sync + 'static> {
     /// by default),
     /// even when this is not the active state,
     /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_fixed_update(&mut self, _data: StateData<'_, T>) {}
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_update(&mut self, _data: StateData<'_, T>) {}
-}
-
-/// An empty `State` trait. It contains no `StateData` or custom `StateEvent`.
-pub trait EmptyState {
-    /// Executed when the game state begins.
-    fn on_start(&mut self, _data: StateData<'_, ()>) {}
-
-    /// Executed when the game state exits.
-    fn on_stop(&mut self, _data: StateData<'_, ()>) {}
-
-    /// Executed when a different game state is pushed onto the stack.
-    fn on_pause(&mut self, _data: StateData<'_, ()>) {}
-
-    /// Executed when the application returns to this game state once again.
-    fn on_resume(&mut self, _data: StateData<'_, ()>) {}
-
-    /// Executed on every frame before updating, for use in reacting to events.
-    fn handle_event(&mut self, _data: StateData<'_, ()>, event: StateEvent) -> EmptyTrans {
-        if let StateEvent::Window(event) = &event {
-            if is_close_requested(&event) {
-                Trans::Quit
-            } else {
-                Trans::None
-            }
-        } else {
-            Trans::None
-        }
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default).
-    fn fixed_update(&mut self, _data: StateData<'_, ()>) -> EmptyTrans {
+    fn shadow_fixed_update(&mut self, _: &mut World) -> Trans<S> {
         Trans::None
     }
 
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit).
-    fn update(&mut self, _data: StateData<'_, ()>) -> EmptyTrans {
+    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
+    /// even when this is not the active state,
+    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
+    fn shadow_update(&mut self, _: &mut World) -> Trans<S> {
         Trans::None
     }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_fixed_update(&mut self, _data: StateData<'_, ()>) {}
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_update(&mut self, _data: StateData<'_, ()>) {}
 }
 
-impl<T: EmptyState> State<(), StateEvent> for T {
-    /// Executed when the game state begins.
-    fn on_start(&mut self, data: StateData<'_, ()>) {
-        self.on_start(data)
-    }
+/// How many callbacks that are inlined.
+const INLINED_CALLBACKS: usize = 16;
 
-    /// Executed when the game state exits.
-    fn on_stop(&mut self, data: StateData<'_, ()>) {
-        self.on_stop(data)
-    }
-
-    /// Executed when a different game state is pushed onto the stack.
-    fn on_pause(&mut self, data: StateData<'_, ()>) {
-        self.on_pause(data)
-    }
-
-    /// Executed when the application returns to this game state once again.
-    fn on_resume(&mut self, data: StateData<'_, ()>) {
-        self.on_resume(data)
-    }
-
-    /// Executed on every frame before updating, for use in reacting to events.
-    fn handle_event(&mut self, data: StateData<'_, ()>, event: StateEvent) -> EmptyTrans {
-        self.handle_event(data, event)
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default).
-    fn fixed_update(&mut self, data: StateData<'_, ()>) -> EmptyTrans {
-        self.fixed_update(data)
-    }
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit).
-    fn update(&mut self, data: StateData<'_, ()>) -> EmptyTrans {
-        self.update(data)
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_fixed_update(&mut self, data: StateData<'_, ()>) {
-        self.shadow_fixed_update(data);
-    }
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_update(&mut self, data: StateData<'_, ()>) {
-        self.shadow_update(data);
-    }
-}
-
-/// A simple `State` trait. It contains `GameData` as its `StateData` and no custom `StateEvent`.
-pub trait SimpleState<'a, 'b> {
-    /// Executed when the game state begins.
-    fn on_start(&mut self, _data: StateData<'_, GameData<'_, '_>>) {}
-
-    /// Executed when the game state exits.
-    fn on_stop(&mut self, _data: StateData<'_, GameData<'_, '_>>) {}
-
-    /// Executed when a different game state is pushed onto the stack.
-    fn on_pause(&mut self, _data: StateData<'_, GameData<'_, '_>>) {}
-
-    /// Executed when the application returns to this game state once again.
-    fn on_resume(&mut self, _data: StateData<'_, GameData<'_, '_>>) {}
-
-    /// Executed on every frame before updating, for use in reacting to events.
-    fn handle_event(
-        &mut self,
-        _data: StateData<'_, GameData<'_, '_>>,
-        event: StateEvent,
-    ) -> SimpleTrans<'a, 'b> {
-        if let StateEvent::Window(event) = &event {
-            if is_close_requested(&event) {
-                Trans::Quit
-            } else {
-                Trans::None
-            }
-        } else {
-            Trans::None
-        }
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default).
-    fn fixed_update(&mut self, _data: StateData<'_, GameData<'_, '_>>) -> SimpleTrans<'a, 'b> {
-        Trans::None
-    }
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit).
-    fn update(&mut self, _data: &mut StateData<'_, GameData<'_, '_>>) -> SimpleTrans<'a, 'b> {
-        Trans::None
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_fixed_update(&mut self, _data: StateData<'_, GameData<'_, '_>>) {}
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_update(&mut self, _data: StateData<'_, GameData<'_, '_>>) {}
-}
-
-impl<'a, 'b, T: SimpleState<'a, 'b>> State<GameData<'a, 'b>, StateEvent> for T {
-    //pub trait SimpleState<'a,'b>: State<GameData<'a,'b>,()> {
-
-    /// Executed when the game state begins.
-    fn on_start(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        self.on_start(data)
-    }
-
-    /// Executed when the game state exits.
-    fn on_stop(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        self.on_stop(data)
-    }
-
-    /// Executed when a different game state is pushed onto the stack.
-    fn on_pause(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        self.on_pause(data)
-    }
-
-    /// Executed when the application returns to this game state once again.
-    fn on_resume(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        self.on_resume(data)
-    }
-
-    /// Executed on every frame before updating, for use in reacting to events.
-    fn handle_event(
-        &mut self,
-        data: StateData<'_, GameData<'_, '_>>,
-        event: StateEvent,
-    ) -> SimpleTrans<'a, 'b> {
-        self.handle_event(data, event)
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default).
-    fn fixed_update(&mut self, data: StateData<'_, GameData<'_, '_>>) -> SimpleTrans<'a, 'b> {
-        self.fixed_update(data)
-    }
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit).
-    fn update(&mut self, mut data: StateData<'_, GameData<'_, '_>>) -> SimpleTrans<'a, 'b> {
-        let r = self.update(&mut data);
-        data.data.update(&data.world);
-        r
-    }
-
-    /// Executed repeatedly at stable, predictable intervals (1/60th of a second
-    /// by default),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_fixed_update(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        self.shadow_fixed_update(data);
-    }
-
-    /// Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit),
-    /// even when this is not the active state,
-    /// as long as this state is on the [StateMachine](struct.StateMachine.html)'s state-stack.
-    fn shadow_update(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        self.shadow_update(data);
-    }
-}
+/// Type alias for a collection of global callbacks.
+type GlobalCallbacks<S, E> = SmallVec<[Box<dyn GlobalCallback<S, E>>; INLINED_CALLBACKS]>;
 
 /// A simple stack-based state machine (pushdown automaton).
 #[derive(Derivative)]
 #[derivative(Debug)]
-pub struct StateMachine<'a, T, E> {
+pub struct StateMachine<S, E>
+where
+    S: State<E>,
+{
     running: bool,
+    /// The stack of the state machine.
     #[derivative(Debug = "ignore")]
-    state_stack: Vec<Box<dyn State<T, E> + 'a>>,
+    stack: Vec<S>,
+    /// Callbacks fired on particular states.
+    #[derivative(Debug = "ignore")]
+    callbacks: S::Storage,
+    /// Callbacks fired on any state.
+    #[derivative(Debug = "ignore")]
+    global_callbacks: GlobalCallbacks<S, E>,
 }
 
-impl<'a, T, E: Send + Sync + 'static> StateMachine<'a, T, E> {
+impl<S, E> StateMachine<S, E>
+where
+    S: State<E>,
+{
     /// Creates a new state machine with the given initial state.
-    pub fn new<S: State<T, E> + 'a>(initial_state: S) -> StateMachine<'a, T, E> {
+    pub fn new() -> StateMachine<S, E> {
         StateMachine {
             running: false,
-            state_stack: vec![Box::new(initial_state)],
+            stack: vec![],
+            callbacks: Default::default(),
+            global_callbacks: Default::default(),
         }
     }
 
+    /// Register a callback associated with a specific state.
+    pub fn register_callback<C: 'static>(&mut self, state: S, callback: C) -> Result<(), Error>
+    where
+        C: StateCallback<S, E>,
+    {
+        self.register_boxed_callback(state, Box::new(callback))
+    }
+
+    /// Register a callback associated with a specific state.
+    pub fn register_boxed_callback(
+        &mut self,
+        state: S,
+        callback: Box<dyn StateCallback<S, E>>,
+    ) -> Result<(), Error> {
+        if self.callbacks.insert(state, callback).is_some() {
+            return Err(Error::StateMachine(StateError::CallbackConflict));
+        }
+
+        Ok(())
+    }
+
+    /// Register a callback associated with a specific state at runtime.
+    ///
+    /// This forcibly overrides any existing states and makes sure that the proper callbacks are
+    /// invoked.
+    pub fn runtime_register_boxed_callback(
+        &mut self,
+        state: S,
+        mut callback: Box<dyn StateCallback<S, E>>,
+        world: &mut World,
+    ) {
+        callback.on_start(world);
+
+        if let Some(mut old) = self.callbacks.insert(state, callback) {
+            old.on_stop(world);
+        }
+    }
+
+    /// Register a global callback that is called on any state.
+    ///
+    /// A global callback received "global" events for this state machine, this includes:
+    ///
+    /// * When it is started.
+    /// * When it is stopped.
+    /// * When it changes state.
+    /// * When it received an event.
+    /// * When it received an update.
+    /// * When it received a fixed update.
+    ///
+    /// This is done through the [`GlobalCallback`](trait.GlobalCallback.html) trait.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # #[macro_use] extern crate amethyst;
+    ///
+    /// use amethyst::{ecs::World, StateMachine, Trans, GlobalCallback};
+    /// use std::{rc::Rc, cell::RefCell};
+    ///
+    /// #[derive(State, Clone, Debug)]
+    /// enum State {
+    ///     First,
+    ///     Second,
+    /// }
+    ///
+    /// #[derive(Debug, Clone, PartialEq, Eq)]
+    /// struct Event(&'static str);
+    ///
+    /// struct Global {
+    ///     capture: Rc<RefCell<Vec<String>>>,
+    /// }
+    ///
+    /// impl GlobalCallback<State, Event> for Global {
+    ///     fn started(&mut self, world: &mut World) {
+    ///         self.capture.borrow_mut().push(format!("Started"));
+    ///     }
+    ///
+    ///     fn stopped(&mut self, world: &mut World) {
+    ///         self.capture.borrow_mut().push(format!("Stopped"));
+    ///     }
+    ///
+    ///     fn changed(&mut self, world: &mut World, state: &State) {
+    ///         self.capture.borrow_mut().push(format!("Changed: {:?}", state));
+    ///     }
+    ///
+    ///     fn handle_event(&mut self, world: &mut World, event: &Event) -> Trans<State> {
+    ///         self.capture.borrow_mut().push(format!("Event: {}", event.0));
+    ///         Trans::None
+    ///     }
+    ///
+    ///     fn update(&mut self, world: &mut World) -> Trans<State> {
+    ///         self.capture.borrow_mut().push(format!("Update"));
+    ///         Trans::None
+    ///     }
+    ///
+    ///     fn fixed_update(&mut self, world: &mut World) -> Trans<State> {
+    ///         self.capture.borrow_mut().push(format!("Fixed Update"));
+    ///         Trans::None
+    ///     }
+    /// }
+    ///
+    /// # fn main() -> amethyst::Result<()> {
+    /// let mut states = StateMachine::new();
+    /// // Helper to capture events.
+    /// let capture = Rc::new(RefCell::new(Vec::new()));
+    /// // The world
+    /// let mut world = World::new();
+    ///
+    /// // Set up the global event handler with our capture.
+    /// states.register_global(Global {
+    ///     capture: Rc::clone(&capture),
+    /// });
+    ///
+    /// states.start(&mut world);
+    /// states.handle_event(&mut world, Event("Message"));
+    ///
+    /// states.update(&mut world);
+    ///
+    /// states.transition(Trans::Switch(State::Second), &mut world);
+    ///
+    /// states.fixed_update(&mut world);
+    ///
+    /// // Note: popping the last state stops the state machine.
+    /// states.transition(Trans::Pop, &mut world);
+    ///
+    /// assert_eq!(
+    ///     *capture.borrow(),
+    ///     &[
+    ///         "Started",
+    ///         "Event: Message",
+    ///         "Update",
+    ///         "Changed: Second",
+    ///         "Fixed Update",
+    ///         "Stopped",
+    ///     ]
+    /// );
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn register_global<C: 'static>(&mut self, callback: C)
+    where
+        C: GlobalCallback<S, E>,
+    {
+        self.register_boxed_global(Box::new(callback));
+    }
+
+    /// Register a boxed global callback that is called on any state.
+    pub fn register_boxed_global(&mut self, callback: Box<dyn GlobalCallback<S, E>>) {
+        self.global_callbacks.push(callback);
+    }
+
     /// Checks whether the state machine is running.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # #[macro_use] extern crate amethyst;
+    /// use amethyst::{ecs::World, StateMachine};
+    ///
+    /// #[derive(State, Clone, Debug)]
+    /// enum State {
+    ///     First,
+    ///     Second,
+    /// }
+    ///
+    /// # fn main() -> amethyst::Result<()> {
+    /// let mut states = StateMachine::<State, ()>::new();
+    /// // the world
+    /// let mut world = World::new();
+    ///
+    /// assert!(!states.is_running());
+    /// states.start(&mut world);
+    /// assert!(states.is_running());
+    /// # Ok(())
+    /// # }
+    /// ```
     pub fn is_running(&self) -> bool {
         self.running
     }
 
     /// Initializes the state machine.
-    pub fn start(&mut self, data: StateData<'_, T>) -> Result<(), StateError> {
-        if !self.running {
-            let state = self
-                .state_stack
-                .last_mut()
-                .ok_or(StateError::NoStatesPresent)?;
-            state.on_start(data);
-            self.running = true;
+    pub fn start(&mut self, world: &mut World) -> Result<(), Error> {
+        if self.running {
+            return Ok(());
         }
+
+        let state = S::default();
+
+        if let Some(c) = self.callbacks.get_mut(&state) {
+            c.on_start(world);
+        }
+
+        for c in &mut self.global_callbacks {
+            c.started(world);
+        }
+
+        self.running = true;
+        self.stack.push(state);
         Ok(())
     }
 
     /// Passes a single event to the active state to handle.
-    pub fn handle_event(&mut self, data: StateData<'_, T>, event: E) {
-        let StateData { world, data } = data;
-        if self.running {
-            let trans = match self.state_stack.last_mut() {
-                Some(state) => state.handle_event(StateData { world, data }, event),
-                None => Trans::None,
-            };
-
-            self.transition(trans, StateData { world, data });
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # #[macro_use] extern crate amethyst;
+    ///
+    /// use amethyst::{ecs::World, StateMachine, Trans, StateCallback};
+    /// use std::{rc::Rc, cell::RefCell};
+    ///
+    /// #[derive(State, Debug, Clone)]
+    /// enum State {
+    ///     First,
+    ///     Second,
+    /// }
+    ///
+    /// #[derive(Debug, Clone, PartialEq, Eq)]
+    /// struct Event(&'static str);
+    ///
+    /// struct SecondState(Rc<RefCell<Option<Event>>>);
+    ///
+    /// impl StateCallback<State, Event> for SecondState {
+    ///     fn handle_event(&mut self, world: &mut World, event: &Event) -> Trans<State> {
+    ///         *self.0.borrow_mut() = Some(event.clone());
+    ///         Trans::None
+    ///     }
+    /// }
+    ///
+    /// # fn main() -> amethyst::Result<()> {
+    /// let mut states = StateMachine::new();
+    /// // Helper to capture the event.
+    /// let capture = Rc::new(RefCell::new(None));
+    /// // The world
+    /// let mut world = World::new();
+    ///
+    /// states.register_callback(State::Second, SecondState(Rc::clone(&capture)))?;
+    /// states.start(&mut world);
+    /// assert_eq!(*capture.borrow(), None);
+    ///
+    /// // Nothing happen because no callback is registered for the initial state.
+    /// states.handle_event(&mut world, Event("hello"));
+    /// assert_eq!(*capture.borrow(), None);
+    ///
+    /// // Transition to second state and verify that we've handled the event.
+    /// states.transition(Trans::Push(State::Second), &mut world);
+    /// states.handle_event(&mut world, Event("world"));
+    /// assert_eq!(*capture.borrow(), Some(Event("world")));
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn handle_event(&mut self, world: &mut World, event: E) {
+        if !self.running {
+            return;
         }
+
+        // Transition which have been requested by callbacks.
+        let mut trans = Trans::None;
+
+        {
+            let StateMachine {
+                ref mut stack,
+                ref mut callbacks,
+                ref mut global_callbacks,
+                ..
+            } = *self;
+
+            for c in global_callbacks {
+                trans.update(c.handle_event(world, &event));
+            }
+
+            if let Some(c) = stack.last().and_then(|s| callbacks.get_mut(s)) {
+                trans.update(c.handle_event(world, &event));
+            }
+        }
+
+        self.transition(trans, world);
     }
 
     /// Updates the currently active state at a steady, fixed interval.
-    pub fn fixed_update(&mut self, data: StateData<'_, T>) {
-        let StateData { world, data } = data;
-        if self.running {
-            let trans = match self.state_stack.last_mut() {
-                Some(state) => state.fixed_update(StateData { world, data }),
-                None => Trans::None,
-            };
-            for state in self.state_stack.iter_mut() {
-                state.shadow_fixed_update(StateData { world, data });
+    pub fn fixed_update(&mut self, world: &mut World) {
+        if !self.running {
+            return;
+        }
+
+        // Transition which have been requested by callbacks.
+        let mut trans = Trans::None;
+
+        {
+            let StateMachine {
+                ref mut stack,
+                ref mut callbacks,
+                ref mut global_callbacks,
+                ..
+            } = *self;
+
+            for c in global_callbacks {
+                trans.update(c.fixed_update(world));
             }
 
-            self.transition(trans, StateData { world, data });
+            // Fixed shadow updates for all states.
+            callbacks.do_values(|c| {
+                trans.update(c.shadow_fixed_update(world));
+            });
+
+            if let Some(c) = stack.last().and_then(|s| callbacks.get_mut(s)) {
+                trans.update(c.fixed_update(world));
+            }
         }
+
+        self.transition(trans, world);
     }
 
     /// Updates the currently active state immediately.
-    pub fn update(&mut self, data: StateData<'_, T>) {
-        let StateData { world, data } = data;
-        if self.running {
-            let trans = match self.state_stack.last_mut() {
-                Some(state) => state.update(StateData { world, data }),
-                None => Trans::None,
-            };
-            for state in self.state_stack.iter_mut() {
-                state.shadow_update(StateData { world, data });
+    pub fn update(&mut self, world: &mut World) {
+        if !self.running {
+            return;
+        }
+
+        // Transition which have been requested by callbacks.
+        let mut trans = Trans::None;
+
+        {
+            let StateMachine {
+                ref mut stack,
+                ref mut callbacks,
+                ref mut global_callbacks,
+                ..
+            } = *self;
+
+            // Regular update for global callbacks.
+            for c in global_callbacks {
+                trans.update(c.update(world));
             }
 
-            self.transition(trans, StateData { world, data });
+            // Shadow updates for all states.
+            callbacks.do_values(|c| {
+                trans.update(c.shadow_update(world));
+            });
+
+            if let Some(c) = stack.last().and_then(|s| callbacks.get_mut(&s)) {
+                trans.update(c.update(world));
+            }
         }
+
+        self.transition(trans, world);
     }
 
     /// Performs a state transition.
-    /// Usually called by update or fixed_update by the user's defined `State`.
-    /// This method can also be called when there are one or multiple `Trans` stored in the
-    /// global `EventChannel<TransEvent<T, E>>`. Such `Trans` will be passed to this method
-    /// sequentially in the order of insertion.
-    pub fn transition(&mut self, request: Trans<T, E>, data: StateData<'_, T>) {
-        if self.running {
-            match request {
-                Trans::None => (),
-                Trans::Pop => self.pop(data),
-                Trans::Push(state) => self.push(state, data),
-                Trans::Switch(state) => self.switch(state, data),
-                Trans::Quit => self.stop(data),
-            }
+    pub fn transition(&mut self, request: Trans<S>, world: &mut World) {
+        if !self.running {
+            return;
+        }
+
+        match request {
+            Trans::None => (),
+            Trans::Pop => self.pop(world),
+            Trans::Push(state) => self.push(state, world),
+            Trans::Switch(state) => self.switch(state, world),
+            Trans::Quit => self.stop(world),
         }
     }
 
     /// Removes the current state on the stack and inserts a different one.
-    fn switch(&mut self, state: Box<dyn State<T, E>>, data: StateData<'_, T>) {
-        if self.running {
-            let StateData { world, data } = data;
-            if let Some(mut state) = self.state_stack.pop() {
-                state.on_stop(StateData { world, data });
-            }
-
-            self.state_stack.push(state);
-
-            //State was just pushed, thus pop will always succeed
-            let state = self.state_stack.last_mut().unwrap();
-            state.on_start(StateData { world, data });
+    fn switch(&mut self, state: S, world: &mut World) {
+        if !self.running {
+            return;
         }
+
+        if let Some(c) = self.stack.pop().and_then(|s| self.callbacks.get_mut(&s)) {
+            c.on_stop(world);
+        }
+
+        if let Some(c) = self.callbacks.get_mut(&state) {
+            c.on_start(world);
+        }
+
+        for c in &mut self.global_callbacks {
+            c.changed(world, &state);
+        }
+
+        self.stack.push(state);
     }
 
     /// Pauses the active state and pushes a new state onto the state stack.
-    fn push(&mut self, state: Box<dyn State<T, E>>, data: StateData<'_, T>) {
-        if self.running {
-            let StateData { world, data } = data;
-            if let Some(state) = self.state_stack.last_mut() {
-                state.on_pause(StateData { world, data });
-            }
-
-            self.state_stack.push(state);
-
-            //State was just pushed, thus pop will always succeed
-            let state = self.state_stack.last_mut().unwrap();
-            state.on_start(StateData { world, data });
+    fn push(&mut self, state: S, world: &mut World) {
+        if !self.running {
+            return;
         }
+
+        let StateMachine {
+            ref mut stack,
+            ref mut callbacks,
+            ..
+        } = *self;
+
+        if let Some(c) = stack.last().and_then(|s| callbacks.get_mut(&s)) {
+            c.on_pause(world);
+        }
+
+        if let Some(c) = callbacks.get_mut(&state) {
+            c.on_start(world);
+        }
+
+        for c in &mut self.global_callbacks {
+            c.changed(world, &state);
+        }
+
+        stack.push(state);
     }
 
     /// Stops and removes the active state and un-pauses the next state on the
     /// stack (if any).
-    fn pop(&mut self, data: StateData<'_, T>) {
-        if self.running {
-            let StateData { world, data } = data;
-            if let Some(mut state) = self.state_stack.pop() {
-                state.on_stop(StateData { world, data });
-            }
+    fn pop(&mut self, world: &mut World) {
+        if !self.running {
+            return;
+        }
 
-            if let Some(state) = self.state_stack.last_mut() {
-                state.on_resume(StateData { world, data });
-            } else {
+        let head = match self.stack.pop() {
+            Some(head) => head,
+            None => return,
+        };
+
+        if let Some(c) = self.callbacks.get_mut(&head) {
+            c.on_stop(world);
+        }
+
+        match self.stack.last() {
+            Some(current) => {
+                if let Some(c) = self.callbacks.get_mut(current) {
+                    c.on_resume(world);
+                }
+
+                for c in &mut self.global_callbacks {
+                    c.changed(world, current);
+                }
+            }
+            None => {
                 self.running = false;
+
+                for c in &mut self.global_callbacks {
+                    c.stopped(world);
+                }
             }
         }
     }
 
     /// Shuts the state machine down.
-    pub(crate) fn stop(&mut self, data: StateData<'_, T>) {
-        if self.running {
-            let StateData { world, data } = data;
-            while let Some(mut state) = self.state_stack.pop() {
-                state.on_stop(StateData { world, data });
-            }
+    pub(crate) fn stop(&mut self, world: &mut World) {
+        if !self.running {
+            return;
+        }
 
-            self.running = false;
+        while let Some(c) = self.stack.pop().and_then(|s| self.callbacks.get_mut(&s)) {
+            c.on_stop(world);
+        }
+
+        for c in &mut self.global_callbacks {
+            c.stopped(world);
+        }
+
+        self.running = false;
+    }
+}
+
+/// Storage implementation for types which only have one value.
+pub struct SingletonStateStorage<S, E> {
+    callback: Option<Box<dyn StateCallback<S, E>>>,
+}
+
+impl<S, E> Default for SingletonStateStorage<S, E> {
+    fn default() -> Self {
+        SingletonStateStorage { callback: None }
+    }
+}
+
+impl<S, E> StateStorage<S, E> for SingletonStateStorage<S, E> {
+    fn insert(
+        &mut self,
+        _: S,
+        callback: Box<dyn StateCallback<S, E>>,
+    ) -> Option<Box<dyn StateCallback<S, E>>> {
+        mem::replace(&mut self.callback, Some(callback))
+    }
+
+    fn get_mut(&mut self, _: &S) -> Option<&mut Box<dyn StateCallback<S, E>>> {
+        self.callback.as_mut()
+    }
+
+    fn do_values<F>(&mut self, mut apply: F)
+    where
+        F: FnMut(&mut Box<dyn StateCallback<S, E>>),
+    {
+        if let Some(c) = self.callback.as_mut() {
+            apply(c);
         }
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+/// Storage implementation for types which can be hashed.
+pub struct MapStateStorage<S, E>
+where
+    S: hash::Hash + PartialEq + Eq,
+{
+    callbacks: hashbrown::HashMap<S, Box<dyn StateCallback<S, E>>>,
+}
 
-    struct State1(u8);
-    struct State2;
-
-    impl State<(), ()> for State1 {
-        fn update(&mut self, _: StateData<'_, ()>) -> Trans<(), ()> {
-            if self.0 > 0 {
-                self.0 -= 1;
-                Trans::None
-            } else {
-                Trans::Switch(Box::new(State2))
-            }
-        }
+impl<S, E> StateStorage<S, E> for MapStateStorage<S, E>
+where
+    S: hash::Hash + PartialEq + Eq,
+{
+    fn insert(
+        &mut self,
+        state: S,
+        callback: Box<dyn StateCallback<S, E>>,
+    ) -> Option<Box<dyn StateCallback<S, E>>> {
+        self.callbacks.insert(state, callback)
     }
 
-    impl State<(), ()> for State2 {
-        fn update(&mut self, _: StateData<'_, ()>) -> Trans<(), ()> {
-            Trans::Pop
-        }
+    fn get_mut(&mut self, state: &S) -> Option<&mut Box<dyn StateCallback<S, E>>> {
+        self.callbacks.get_mut(state)
     }
 
-    #[test]
-    fn switch_pop() {
-        use crate::ecs::prelude::World;
-
-        let mut world = World::new();
-
-        let mut sm = StateMachine::new(State1(7));
-        // Unwrap here is fine because start can only fail when there are no states in the machine.
-        sm.start(StateData::new(&mut world, &mut ())).unwrap();
-
-        for _ in 0..8 {
-            sm.update(StateData::new(&mut world, &mut ()));
-            assert!(sm.is_running());
+    fn do_values<F>(&mut self, mut apply: F)
+    where
+        F: FnMut(&mut Box<dyn StateCallback<S, E>>),
+    {
+        for c in self.callbacks.values_mut() {
+            apply(c);
         }
-
-        sm.update(StateData::new(&mut world, &mut ()));
-        assert!(!sm.is_running());
     }
 }
+
+impl<S, E> Default for MapStateStorage<S, E>
+where
+    S: hash::Hash + PartialEq + Eq,
+{
+    fn default() -> Self {
+        MapStateStorage {
+            callbacks: hashbrown::HashMap::new(),
+        }
+    }
+}
+
+impl<E> State<E> for () {
+    type Storage = SingletonStateStorage<Self, E>;
+}
+
+/// Helper macro to implement storage for certain types.
+macro_rules! impl_map_storage {
+    ($lt:lifetime, $ty:ty) => {
+        impl<$lt, E> State<E> for &$lt $ty {
+            type Storage = MapStateStorage<Self, E>;
+        }
+    };
+
+    ($ty:ty) => {
+        impl<E> State<E> for $ty {
+            type Storage = MapStateStorage<Self, E>;
+        }
+    };
+}
+
+/// Types that can be used as states through `MapStateStorage`.
+impl_map_storage!('a, str);
+impl_map_storage!(u8);
+impl_map_storage!(u32);
+impl_map_storage!(u64);
+impl_map_storage!(i8);
+impl_map_storage!(i32);
+impl_map_storage!(i64);
